Refactor Gui
