+++ src/main/java/lania/Gui.java
@@ -1,0 +1,123 @@
+package lania;
+
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.Region;
+import javafx.scene.layout.VBox;
+import javafx.stage.Stage;
+import lania.command.Command;
+import lania.exception.LaniaException;
+import lania.task.TaskList;
+
+import java.time.format.DateTimeParseException;
+
+public class Gui {
+
+    private Stage stage;
+    private TaskList tasks;
+    private Ui ui;
+    private Storage storage;
+    private Parser parser;
+
+    private ScrollPane scrollPane;
+    private VBox dialogContainer;
+    private TextField userInput;
+    private Button sendButton;
+    private Scene scene;
+    private Image user = new Image(this.getClass().getResourceAsStream("/images/User.png"));
+    private Image duke = new Image(this.getClass().getResourceAsStream("/images/Lania.png"));
+
+    public Gui (Stage stage, TaskList tasks, Ui ui, Storage storage, Parser parser) {
+        this.stage = stage;
+        this.tasks = tasks;
+        this.ui = ui;
+        this.storage = storage;
+        this.parser = parser;
+    }
+
+    public void runGui() {
+        scrollPane = new ScrollPane();
+        dialogContainer = new VBox();
+        scrollPane.setContent(dialogContainer);
+
+        userInput = new TextField();
+        sendButton = new Button("Send");
+
+        AnchorPane mainLayout = new AnchorPane();
+        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
+
+        scene = new Scene(mainLayout);
+
+        stage.setScene(scene);
+        stage.show();
+        stage.setTitle("Lania");
+        stage.setResizable(false);
+        stage.setMinHeight(600.0);
+        stage.setMinWidth(400.0);
+
+        mainLayout.setPrefSize(400.0, 600.0);
+
+        scrollPane.setPrefSize(385, 535);
+        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
+        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
+
+        scrollPane.setVvalue(1.0);
+        scrollPane.setFitToWidth(true);
+
+        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
+
+        userInput.setPrefWidth(325.0);
+
+        sendButton.setPrefWidth(55.0);
+
+        AnchorPane.setTopAnchor(scrollPane, 1.0);
+
+        AnchorPane.setBottomAnchor(sendButton, 1.0);
+        AnchorPane.setRightAnchor(sendButton, 1.0);
+
+        AnchorPane.setLeftAnchor(userInput , 1.0);
+        AnchorPane.setBottomAnchor(userInput, 1.0);
+
+        dialogContainer.getChildren().addAll(
+                DialogBox.getLaniaDialog(new Label(ui.showGreetingMessage()), new ImageView(duke)));
+
+        sendButton.setOnMouseClicked((event) -> {
+            handleUserInput();
+        });
+
+        userInput.setOnAction((event) -> {
+            handleUserInput();
+        });
+
+        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
+    }
+
+    private void handleUserInput() {
+        Label userText = new Label(userInput.getText());
+        try {
+            Command c = parser.parse(userInput.getText());
+            Label dukeText = new Label(c.execute(tasks, storage, ui));
+            dialogContainer.getChildren().addAll(
+                    DialogBox.getUserDialog(userText, new ImageView(user)),
+                    DialogBox.getLaniaDialog(dukeText, new ImageView(duke))
+            );
+        } catch (LaniaException e) {
+            dialogContainer.getChildren().addAll(
+                    DialogBox.getUserDialog(userText, new ImageView(user)),
+                    DialogBox.getLaniaDialog(new Label(ui.showLaniaException(e)), new ImageView(duke))
+            );
+        } catch (DateTimeParseException e) {
+            dialogContainer.getChildren().addAll(
+                    DialogBox.getUserDialog(userText, new ImageView(user)),
+                    DialogBox.getLaniaDialog(new Label(ui.showDateTimeException()), new ImageView(duke))
+            );
+        }
+        userInput.clear();
+    }
+}
